window: Be smarter about sizing windows
If width-for-height comes up with a too large natural width for
its height, try height-for-width, too. If that can fit its natural
height, prefer it.

This way, windows that care about aspect ratio, but where the monitor is
landscape and the window is portrait (or really: just a smaller aspect
ratio than the monitor) can end up with the correct aspect
ratio