window: Don't make initial window size 100% of screen size
Take 85%, because it's the same magic number eog uses, so that
windows don't hug the edge of the screen and look like they're
maximized.

We only do this for the first dimension because it's better to match
aspect ratios than it is to not hug the screen.

Note that we do not do that if ::default-size is set because an
explicitly set width means somebody knows what they're doing.