Add GtkGlyphPaintable
This is a paintable that wraps a font file,
and renders a single glyph. The glyph can be
selected by the "glyph" property, but the
paintable looks for a glyph with the name
"icon0" first, so single-icon fonts can be
made to work out of the box.

The paintable has properties to set font
variations, as well as colors.

It implements GtkSymbolicPaintable, and will
override the foreground, error, warning and
success colors with theme colors when used
symbolically.