GStreamer Media backend: Pass WGL context ownership back to GdkGL
We have no good way to know whether GStreamer is able to playback the file that
was fed into it, and when it couldn't (for instance, the installed plugins
cannot support decoding the file format), the control of the underlying WGL
context is in the GstGL's thread, not GTK's own thread.

As a result, we can't just do GdkGLContext calls directly in such scenarios,
otherwise the program crashes as the function pointers held by libepoxy (used
by GdkGLContext) are rendered invalid.  So, to be safe, if OpenGL (WGL) is
used, call epoxy_handle_external_wglMakeCurrent() upon sink disposal.

Fixes issue #5685.