GDK-Win32: Backport MR !5702 into gtk-3-24
Since we might be using GdkGLContext across different threads and may be using
GdkGLContext with different libraries or programs that make the WGL context
current outside of GTK/GDK, make things safer for such applications or
libraries, such as GstGL.

This will make some wgl*() calls in ->make_current() and ->dispose() link
directly to the system's/ICD opengl32.dll instead of going through the function
pointers that are acquired via libepoxy, which can be rendered invalid when the
WGL context is made current outside of GdkGLContext, leading to weird crashes.
By linking directly to opengl32.dll, we are assured that these wgl*() calls
will be carried out correctly.

Note that this is not done when initializing or realizing the WGL context as
we are in the same thread as these are all carried out within GTK/GDK, as it
should not be necessary to do so there.