wayland: Remove an old workaround
This workaround - were it ever to trigger - is broken today. It destroys
the wl_surface and all associated structs but does not recreate the
xdg_popup or xdg_toplevel struct, so it would cause a hidden window.

The workaround looked a lot different when it was introduced in commit
83b54bab5734f17d01633c1b5a0ce3e4896d2ee0, too - both in what it did and
in what the vfuncs did that it called.